requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile
#
attrs==21.4.0
    # via fiona
certifi==2022.5.18.1
    # via
    #   fiona
    #   pyproj
click==8.1.3
    # via fiona
click-plugins==1.1.1
    # via fiona
cligj==0.7.2
    # via fiona
fiona==1.8.21
    # via geopandas
geopandas==0.10.2
    # via frechet (setup.py)
munch==2.5.0
    # via fiona
numpy==1.22.4
    # via pandas
pandas==1.4.2
    # via
    #   frechet (setup.py)
    #   geopandas
pyproj==3.3.1
    # via geopandas
python-dateutil==2.8.2
    # via pandas
python-dotenv==0.20.0
    # via frechet (setup.py)
pytz==2022.1
    # via pandas
shapely==1.8.2
    # via geopandas
six==1.16.0
    # via
    #   fiona
    #   python-dateutil

# The following packages are considered to be unsafe in a requirements file:
# setuptools
